src/app/api/recommender.py
import pandas as pd
from pydantic import BaseModel
from .data_loader import books, db_books

class RecommendationRequest(BaseModel):
    query: str
    category: str = "All"
    tone: str = "All"


def retrieve_semantic_recommendations(
    query: str,
    category: str = None,
    tone: str = None,
    initial_top_k: int = 50,
    final_top_k: int = 16,
) -> pd.DataFrame:
    """
    Retrieves and filters book recommendations based on semantic similarity and user-selected criteria.
    """
    recs = db_books.similarity_search(query, k=initial_top_k)
    books_list = [int(rec.page_content.strip('"').split()[0]) for rec in recs]
    book_recs = books[books["isbn13"].isin(books_list)].head(initial_top_k)

    if category != "All":
        book_recs = book_recs[book_recs["simple_categories"]
                              == category].head(final_top_k)
    else:
        book_recs = book_recs.head(final_top_k)

    if tone == "Happy":
        book_recs.sort_values(by="joy", ascending=False, inplace=True)
    elif tone == "Surprising":
        book_recs.sort_values(by="surprise", ascending=False, inplace=True)
    elif tone == "Angry":
        book_recs.sort_values(by="anger", ascending=False, inplace=True)
    elif tone == "Suspenseful":
        book_recs.sort_values(by="fear", ascending=False, inplace=True)
    elif tone == "Sad":
        book_recs.sort_values(by="sadness", ascending=False, inplace=True)

    return book_recs


def format_recommendations(recommendations: pd.DataFrame):
    """
    Formats the DataFrame of recommendations into a list of dictionaries for the API response.
    """
    results = []
    for _, row in recommendations.iterrows():
        description = row["description"]
        truncated_desc_split = description.split()
        truncated_description = " ".join(truncated_desc_split[:30]) + "..."

        authors_split = row["authors"].split(";")
        if len(authors_split) == 2:
            authors_str = f"{authors_split[0]} and {authors_split[1]}"
        elif len(authors_split) > 2:
            authors_str = f"{', '.join(authors_split[:-1])}, and {authors_split[-1]}"
        else:
            authors_str = row["authors"]

        results.append({
            "title": row["title"],
            "authors": authors_str,
            "thumbnail": row["large_thumbnail"],
            "description": truncated_description,
        })
    return results


def get_recommendations_logic(query: str, category: str, tone: str):
    """
    Main logic to retrieve and format recommendations.
    """
    recommendations = retrieve_semantic_recommendations(query, category, tone)
    return format_recommendations(recommendations)
