pyproject.toml
[project]
name = "beyond-the-bookshelf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "fastapi[all,standard]>=0.116.1",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.31",
    "langchain-text-splitters>=0.3.9",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.55.3",
    "uvicorn[standard]>=0.35.0",
]
